backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.8.2
openai==1.51.0
langgraph==0.2.35
langchain-groq==0.2.0
langchain-core==0.3.15
apscheduler==3.10.4
twilio==9.3.2
python-dotenv==1.0.1
httpx==0.27.2
